feat: Java testgen class name fix, remove per-test @Timeout, and wire language_version
- Add class_name and qualified_name to /testgen API payload so the backend
  has explicit access to computed FunctionToOptimize properties
- Add client-side _fix_java_test_class_name() to correct wrong class name
  references in LLM-generated Java test code
- Remove per-test @Timeout annotation from Java instrumentation (causes
  timing instability on CI runners; Maven Surefire handles timeouts)
- Remove redundant default_language_version, use language_version as canonical

+++ tests/test_languages/test_java/test_instrumentation.py
@@ -221,6 +221,7 @@ public class FibonacciTest {
     }
 }
 """
+
         test_file.write_text(source)
 
         func = FunctionToOptimize(
@@ -2704,7 +2705,7 @@ public class CounterTest__perfinstrumented {
                 }
             }
         }
-        assertEquals(1, _cf_result1_1);
+        assertEquals(1, (int)_cf_result1_1);
     }
 }
 """
